backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
boto3==1.10.28
botocore==1.13.29         # via boto3, s3transfer
click==7.0
dj-database-url==0.5.0
django-configurations==2.1
django-cors-headers==3.2.0
django-extensions==2.2.5
django==2.2.7
djangorestframework==3.10.3
docutils==0.15.2          # via botocore
jinja2==2.10.3
jmespath==0.9.4           # via boto3, botocore
markupsafe==1.1.1
pip-tools==4.3.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0    # via botocore
pytz==2019.3
s3transfer==0.2.1         # via boto3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7           # via botocore
